% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{H}
\alias{H}
\alias{G}
\title{Copula H and G functions}
\usage{
H(u, v, copula = "normal", param = list())

G(u, w, copula = "normal", param = list())
}
\arguments{
\item{u}{latent uniform(0,1) variable.}

\item{v}{uniform(0,1) errors.}

\item{copula}{model (see 'Details').}

\item{param}{\code{list} of parameters to pass to the model.}

\item{w}{uniform(0,1) observable variables}
}
\value{
A vector or matrix of the same dimension as \code{v} for \code{H}
or as \code{w} for \code{G}.
}
\description{
Copula H and G functions
}
\details{
Currently we support the following \code{model}: 
"\code{normal}", "\code{student}", "\code{skewt}" (skew-t distribution),
"\code{ghyp}" (generalized hyperbolic distribution), "\code{clayton}", "\code{gumbel}".
The list of \code{param} depends on the model:
\itemize{
   \item \code{normal}: either correlation \code{corr} (between -1 and 1) or 
   variances of \eqn{Z} \code{var_z}\eqn{>0} and \eqn{\epsilon} \code{var_e}\eqn{>0};
   \item \code{student}: degrees of freedom \code{nu}\eqn{>0};
   \item \code{skewt}: degrees of freedom \code{nu}\eqn{>0}, 
   the variance of \eqn{\epsilon} \code{var_e}\eqn{>0} and 
   the parameter of asymetry \code{lambda} (betweem -1 and 1);
   \item \code{ghyp}: \code{lambda}, \code{chi} (positive), \code{psi} (positive),
   the variance of \eqn{\epsilon} \code{var_e}\eqn{>0} and \code{gamma};
   \item \code{clayton}: \code{alpha} (positive), the parameter of dependence;
   \item \code{gumbel}: \code{alpha}\eqn{>1}, the parameter of dependence.
}
}
