% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{lp_fit_Spline}
\alias{lp_fit_Spline}
\title{Linear program to estimate spline coefficients}
\usage{
lp_fit_Spline(
  u,
  v,
  w,
  G = TRUE,
  type = "b",
  splines_control = splines.control()
)
}
\arguments{
\item{u}{vector of latent uniform(0,1) variable.}

\item{v}{uniform(0,1) vector of error.}

\item{w}{uniform(0,1) vector of observation.}

\item{G}{if \code{TRUE} estimate function \code{G} (\code{\link{G}}), 
otherwise estimate function \code{H} (\code{\link{H}}) (see 'Details').}

\item{type}{specify spline basis, either \code{"b"} (default), 
\code{"c"}, \code{"i"} or \code{"m"};}

\item{splines_control}{control (see \code{\link{splines.control}}).}
}
\value{
A matrix of spline coefficients
}
\description{
Estimate the spline coefficients of the tensor product spline
by minimising
\deqn{\sum_{i=1}^N|x_i-\phi(u)^TX\psi(y_i)|,}
under monotonicity constraint (nonnegative partial derivatives).
}
