% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{map}
\alias{map}
\title{Maximum a posteriori (MAP) estimator}
\usage{
map(x, w, type = "b", splines_control = splines.control())
}
\arguments{
\item{x}{vector spline coefficients (\eqn{\lambda} vectorized).}

\item{w}{uniform(0,1) \code{matrix} of observations.}

\item{type}{specify spline basis, either \code{"b"} (default), 
\code{"c"}, \code{"i"} or \code{"m"};}

\item{splines_control}{control (see \code{\link{splines.control}}).}
}
\value{
Vector of estimates (MAP) for \eqn{u}
}
\description{
Estimate the latent variable \eqn{u}.
}
