% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{mle.control}
\alias{mle.control}
\title{Control for fitting the MLE}
\usage{
mle.control(
  stopval = -Inf,
  xtol_rel = 1e-06,
  maxeval = 30000,
  ftol_rel = 0,
  ftol_abs = 0,
  check_derivatives = FALSE
)
}
\arguments{
\item{stopval}{stop minimization at this value}

\item{xtol_rel}{tolerance for step}

\item{maxeval}{max number of function evaluations (default is 30000)}

\item{ftol_rel}{relative tolerance}

\item{ftol_abs}{absolute tolerance}
}
\description{
Control for the optimization routine
for the MLE (see \code{\link{nfcm_mle}}). Currently the
optimization routine is a sequential quadratic programming
(SQP) algorithm \code{\link[nloptr]{slsqp}} from the \code{nloptr} 
package, and thus the control are the same as \code{\link[nloptr]{nl.opts}}
}
\seealso{
\code{\link[nloptr]{nl.opts}}, \code{\link[nloptr]{slsqp}}
}
