% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{nfcm_mle}
\alias{nfcm_mle}
\title{Maximum likelihood estimator of Spline Based Nonlinear Factor Copula Model}
\usage{
nfcm_mle(
  lambda,
  w,
  type = "b",
  splines_control = splines.control(),
  mle_control = mle.control()
)
}
\arguments{
\item{lambda}{vector of starting values for spline coefficients (\eqn{\Lambda} vectorized).}

\item{w}{uniform(0,1) \eqn{n\times N} \code{matrix} of observations.}

\item{type}{specify spline basis, either \code{"b"} (default), 
\code{"c"}, \code{"i"} or \code{"m"};}

\item{splines_control}{control (see \code{\link{splines.control}}).}

\item{mle_control}{control to pass to the optimization routine (see 'Details').}
}
\value{
An object from the optimization routine (see 'Details')
}
\description{
Minimize the negative log-likelihood for a bivariate model. The copula models assume
positive quadrant dependent variables. If one has \eqn{N}-dimensional
variable, the variable is transformed to a 2-dimensional matrix (see 'Details').
}
\seealso{
\code{\link[nloptr]{slsqp}}
}
