% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{P}
\alias{P}
\title{"P" matrix}
\usage{
P(type = "b", splines_control = splines.control())
}
\arguments{
\item{type}{specify spline basis, either \code{"b"} (default), 
\code{"c"}, \code{"i"} or \code{"m"};}

\item{splines_control}{control (see \code{\link{splines.control}}).}
}
\description{
Compute the "P" matrix, which is defined as 
\deqn{\int_0^1\phi(u)\phi(u)^T d u},
where \eqn{\phi} is a spline basis.
}
\details{
Integral is approximated using \code{\link[stats]{integrate}}.
}
