% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot.H.spline}
\alias{plot.H.spline}
\alias{plot.G.spline}
\title{Plotting the function H or G approximated by splines}
\usage{
\method{plot}{H.spline}(x, n = 100, ...)

\method{plot}{G.spline}(x, n = 100, ...)
}
\description{
Plotting the function H or G approximated by splines
}
