% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{nfcm_mle}
\alias{nfcm_mle}
\title{Maximum likelihood estimator of Spline Based Nonlinear Factor Copula Model}
\usage{
nfcm_mle(
  x,
  w1,
  w2 = NULL,
  P = NULL,
  type = "b",
  splines_control = splines.control(),
  mle_control = mle.control()
)
}
\arguments{
\item{x}{vector of starting values for spline coefficients (\eqn{\lambda} vectorized).}

\item{w1}{uniform(0,1) vector of observations, or \code{matrix}
if \code{w2} is \code{NULL}.}

\item{w2}{uniform(0,1) vector of observations, can be set to \code{NULL} (by default)
if \code{w1} is a \code{matrix}.}

\item{P}{the "P" matrix (see \code{\link{P}}), if \code{NULL} (by default) it is 
computed when calling the function.}

\item{type}{specify spline basis, either \code{"b"} (default), 
\code{"c"}, \code{"i"} or \code{"m"};}

\item{splines_control}{control (see \code{\link{splines.control}}).}

\item{fit_control}{control to pass to the optimization routine (see 'Details').}
}
\value{
An object from the optimization routine (see 'Details')
}
\description{
Minimize the negative log-likelihood for a bivariate model. The copula models assume
positive quadrant dependent variables. If one has \eqn{N}-dimensional
variable, the variable is transformed to a 2-dimensional matrix (see 'Details').
}
\seealso{
\code{\link[nloptr]slsqp}
}
