% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot.H.spline}
\alias{plot.H.spline}
\alias{plot.G.spline}
\title{Plotting the function H or G approximated by splines}
\usage{
\method{plot}{H.spline}(x, n = 100, ...)

\method{plot}{G.spline}(x, n = 100, ...)
}
\arguments{
\item{x}{an object of class 'H' or 'G' (see 'nfcm_mle')}

\item{n}{number of points to evaluate the function}

\item{...}{additional arguments to be passed to 'persp'}
}
\description{
Plotting the function H or G approximated by splines
}
\seealso{
\code{\link[graphics]{persp}}
}
