% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{predict_u}
\alias{predict_u}
\title{Predict latent variable}
\usage{
predict_u(
  lambda,
  w,
  type = "b",
  splines_control = splines.control(),
  method = "map"
)
}
\arguments{
\item{lambda}{vector spline coefficients (\eqn{\Lambda} vectorized).}

\item{w}{uniform(0,1) \code{matrix} of observations.}

\item{type}{specify spline basis, either \code{"b"} (default), 
\code{"c"}, \code{"i"} or \code{"m"};}

\item{splines_control}{control (see \code{\link{splines.control}}).}

\item{method}{method to estimate the latent variable, either \code{"map"} (default) or \code{"blup"}.}
}
\value{
Estimates for \eqn{u}, return a vector if \code{method="map"} and a matrix if \code{method="blup"}.
}
\description{
Estimate the latent variable \eqn{u}.
}
